% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_contour}
\alias{create_contour}
\title{Plot a power contour}
\usage{
create_contour(
  x,
  y,
  pwr,
  pwrcolours = seq(0, 1, by = 0.1),
  xlab = "Sample size",
  ylab = "Hypothetical effect size",
  pwrbands = c(0.5, 0.8, 0.95),
  alpha = 0.6
)
}
\arguments{
\item{x}{The x values.}

\item{y}{The y values.}

\item{pwr}{The power values.}

\item{pwrcolours}{The breaks for the power bands. Defaults to \code{seq(0,1,0.1)}.}

\item{xlab}{The label of the x axis.}

\item{ylab}{The label of the y axis.}

\item{pwrbands}{The additional black power bands. Defaults to
\code{c(0.5, 0.8, 0.95)}.}

\item{alpha}{The transparency of the contour for seeing the grid lines
underneath.}
}
\value{
A \code{ggplot2} object.
}
\description{
This is the function which creates the contours from a set of values. These
can either be generated with the contourcalc functions, or with any custom
values derived from, for example, simulations.
}
\examples{
vals <- pwrcontour:::pwr.t.test_contourcalc(5, 50, 0.3, 0.7)
create_contour(vals$n, vals$ES, vals$power, xlab="x", ylab="y")
}
