% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{pwr.eq.r.test.table}
\alias{pwr.eq.r.test.table}
\title{Create a Power Table for a for a test of equivalence for a correlation}
\usage{
pwr.eq.r.test.table(n, sig.level = 0.05, kable = TRUE, digits = 3, ...)
}
\arguments{
\item{n}{The proposed sample size.}

\item{sig.level}{Significance level, or alpha. Defaults to 0.05.}

\item{kable}{Should the table be presented as a \code{knitr::kable()} figure? Defaults to \code{TRUE}. Otherwise a \code{tibble} is returned.}

\item{digits}{The number of digits.}

\item{...}{Extra input arguments for the \code{knitr::kable()} function}
}
\value{
A power table.
}
\description{
Plot a power analysis table of effect size and power for Pearson's r tests
in the style of Richard Morey's jpower JAMOVI module,
assuming the true effect is equal to zero.
}
\examples{
pwr.r.test.table(n=25)
}
