% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{pwr.eq.t.test.contour}
\alias{pwr.eq.t.test.contour}
\title{Plot a Power Contour for a test of equivalence for a t-test}
\usage{
pwr.eq.t.test.contour(
  nmin,
  nmax,
  esmin,
  esmax,
  sig.level = 0.05,
  type = c("two.sample", "one.sample", "paired"),
  pwrbands = c(0.5, 0.8, 0.95),
  d_unitconversion = 1,
  d_units = "Cohen's d",
  alpha = 0.6
)
}
\arguments{
\item{nmin}{Minimum sample size for the plot.}

\item{nmax}{Maximum sample size for the plot.}

\item{esmin}{Minimum effect size for the plot. Defaults to Cohen's d, but can be percentages or units of measurement if desired using the \code{d_unitconversion} parameter.}

\item{esmax}{Maximum effect size for the plot. Defaults to Cohen's d, but can be percentages or units of measurement if desired using the \code{d_unitconversion} parameter.}

\item{sig.level}{Significance level, or alpha. Defaults to 0.05.}

\item{type}{The t-test type. The default is "two.sample", but other options are "paired" or "one.sample".}

\item{pwrbands}{Additional black power bands to display on the figure to emphasise certain power levels. Defaults to c(0.5, 0.8, 0.95).}

\item{d_unitconversion}{Unit conversion of Cohen's d. If specified as 5, for example, 5 units of change in the units of the figure are considered equivalent to Cohen's d = 1. This conversion should be calculated in advance. Defaults to 1 for Cohen's d.}

\item{d_units}{The units of the difference. Defaults to "Cohen's d".}

\item{alpha}{The transparency of the contours, in order to show grid lines below.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a contour of sample size, effect size and power for t-test
of equivalence in the style of Richard Morey's jpower JAMOVI module,
assuming the true effect is equal to zero.
}
\examples{
# A straightforward example
pwr.eq.t.test.contour(3, 100, 0.1, 1, sig.level=0.05)

# Custom units
pwr.eq.t.test.contour(3, 100, 0.5, 5, sig.level=0.05,
                   d_unitconversion=5, d_units="mmHg")

# Adding things
pwr.eq.t.test.contour(3, 100, 0.5, 5, sig.level=0.05,
                   d_unitconversion=5, d_units="mmHg") +
                   ggplot2::annotate("text", x=25, y=2,
                   label="We don't want to be here")
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
