% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{pwr.prop.test.contour}
\alias{pwr.prop.test.contour}
\title{Plot a Power Contour for a proportion test}
\usage{
pwr.prop.test.contour(
  nmin,
  nmax,
  propmin,
  propmax,
  propcomp = 0.5,
  sig.level = 0.05,
  type = c("two.sample", "one.sample"),
  alternative = c("two.sided", "less", "greater"),
  pwrbands = c(0.5, 0.8, 0.95),
  alpha = 0.6
)
}
\arguments{
\item{nmin}{Minimum sample size for the plot.}

\item{nmax}{Maximum sample size for the plot.}

\item{propmin}{Minimum proportion for the plot.}

\item{propmax}{Maximum proportion for the plot.}

\item{propcomp}{Comparison proportion value against which the measured proportion is compared.}

\item{sig.level}{Significance level, or alpha. Defaults to 0.05.}

\item{type}{The proportion test type. The default is "two.sample", but another option is "one.sample".}

\item{alternative}{The alternative hypothesis. The default is "two.sided", but other options are "less" or "greater".}

\item{pwrbands}{Additional black power bands to display on the figure to emphasise certain power levels. Defaults to c(0.5, 0.8, 0.95).}

\item{alpha}{The transparency of the contours, in order to show grid lines below.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a contour of sample size, effect size and power for proportion test
comparisons in the style of Richard Morey's jpower JAMOVI module.
}
\examples{
pwr.prop.test.contour(5, 100, 0.1, 0.9)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
