% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{pwr.r.test.contour}
\alias{pwr.r.test.contour}
\title{Plot a Power Contour for a correlation test comparison}
\usage{
pwr.r.test.contour(
  nmin,
  nmax,
  rmin,
  rmax,
  sig.level = 0.05,
  alternative = c("two.sided", "less", "greater"),
  pwrbands = c(0.5, 0.8, 0.95),
  alpha = 0.6
)
}
\arguments{
\item{nmin}{Minimum sample size for the plot.}

\item{nmax}{Maximum sample size for the plot.}

\item{rmin}{Minimum Pearson's r for the plot.}

\item{rmax}{Maximum Pearson's r for the plot.}

\item{sig.level}{Significance level, or alpha. Defaults to 0.05.}

\item{alternative}{The alternative hypothesis. The default is "two.sided", but other options are "less" or "greater".}

\item{pwrbands}{Additional black power bands to display on the figure to emphasise certain power levels. Defaults to c(0.5, 0.8, 0.95).}

\item{alpha}{The transparency of the contours, in order to show grid lines below.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a contour of sample size, effect size and power for correlation comparisons
in the style of Richard Morey's jpower JAMOVI module.
}
\examples{
pwr.r.test.contour(5, 100, 0.1, 0.7, sig.level=0.05)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
