% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{pwr.t.test.table}
\alias{pwr.t.test.table}
\title{Create a Power Table for a t-test comparison}
\usage{
pwr.t.test.table(
  n,
  sig.level = 0.05,
  type = c("two.sample", "one.sample", "paired"),
  alternative = c("two.sided", "less", "greater"),
  d_unitconversion = 1,
  d_units = "Cohen's d",
  kable = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{n}{The proposed sample size.}

\item{sig.level}{Significance level, or alpha. Defaults to 0.05.}

\item{type}{The t-test type. The default is "two.sample", but other options are "paired" or "one.sample".}

\item{alternative}{The alternative hypothesis. The default is "two.sided", but other options are "less" or "greater".}

\item{d_unitconversion}{Unit conversion of Cohen's d. If specified as 5, for example, 5 units of change in the units of the table are considered equivalent to Cohen's d = 1. This conversion should be calculated in advance. Defaults to 1 for Cohen's d.}

\item{d_units}{The units of the difference. Defaults to "Cohen's d".}

\item{kable}{Should the table be presented as a \code{knitr::kable()} figure? Defaults to \code{TRUE}. Otherwise a \code{tibble} is returned.}

\item{digits}{The number of digits.}

\item{...}{Extra input arguments for the \code{knitr::kable()} function}
}
\value{
A power table.
}
\description{
Plot a power analysis table of effect size and power for t-test comparisons
in the style of Richard Morey's jpower JAMOVI module.
}
\examples{
# Straightforward example
pwr.t.test.table(n=25)

# More customised example
pwr.t.test.table(n=25, d_unitconversion=20, d_units="\%", digits=1)
}
\author{
Granville J Matheson, \email{mathesong@gmail.com}
}
